Add functions to clear and add numbers to screen.
The functions for this commit work such that
when a button is clicked the object, objectInfo
is created and separate pieces of information are
sent to functions to determine the proper behavior
that should take place. This object contains
the following:

•Target of the click event
•ClassList => Array
•textContent of the button that called the event.

These 3 bits information aim to capture the
appropriate information for the logic.